Specify time trends: age pattern fixed

Specify time trends where the shape of the age profile is assumed to be constant over time, even though the level may change. The level is assumed to follow a random walk.

Usage

TimeFixed(scale = 1)

Arguments

scale

Scale for the standard deviation of the error term. Default is 1.

Value

An object of class "BayesRates_spec_timefixed".

Mathematical details

In the model underlying TimeFixed(),

$$\eta_{at} = \alpha_a + \beta_t$$

where

\(\eta_{at}\) denotes (log) rates for age group \(a\) during period \(t\),

\(\alpha_a\) is an age effect,

\(\beta_t\) is a time effect,

and the time effect is modelled as

$$\beta_t = \beta_{t-1} + \epsilon_t$$

where \(\epsilon\) is a normally-distributed error term with mean \(0\) and standard deviation \(\sigma\).

Standard deviation \(\sigma\) has prior

$$\sigma \sim \text{N}^+(0, a^2)$$

where \(\text{N}^+\) denotes a half-normal distribution, and a value for scale parameter \(a\) is supplied by the user.

See also

Examples

TimeFixed()
#> < Object of class "BayesRates_spec_timefixed" >
#>           scale: 1
TimeFixed(scale = 2)
#> < Object of class "BayesRates_spec_timefixed" >
#>           scale: 2