Obtain estimates of rates, combined with original data

Extract estimated rates from an object created by a call to smooth_age() or smooth_agetime(), and combine these with the original data.

Usage

# S3 method for class 'BayesRates_results'
augment(x, interval = 0.95, ...)

Arguments

x

A "BayesRates_results" object created by calling function smooth_age() or smooth_agetime().

interval

Width of credible intervals. A number between 0 and 1. Default is 0.95.

...

Not currently used.

Value

A tibble, constructed by merging nevent_df and py_df, and the creating six new columns:

<agevar>.mid: Midpoints of age groups (which is useful for plotting.)

.fitted: Point estimates (posterior medians) of rates.

.lower, .upper: Lower and upper bounds of credible intervals specified by interval.

.probability: A list column with all draws from the posterior distribution.

.observed: Direct estimates of rates.

Details

In addition to the modelled rates, augment() also displayed 'direct' estimates of rates. Direct estimates are estimates obtained by dividing events by populations at risk, with no smoothing.

See also

Examples

results <- smooth_agetime(nevent_df = nz_divorces,
                          py_df = nz_population,
                          age_width_df = nz_age_width_df,
                          age_min = 15)
augment(results)
#> # A tibble: 660 × 11
#>    age    time sex    nevent     py age.mid .fitted  .lower  .upper .probability
#>    <fct> <int> <chr>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <list>      
#>  1 15-19  1992 Female     12 139500    17.5 2.97e-5 2.46e-5 3.61e-5 <dbl>       
#>  2 15-19  1992 Male        0 140660    17.5 2.97e-5 2.46e-5 3.61e-5 <dbl>       
#>  3 15-19  1993 Female      6 135510    17.5 2.84e-5 2.31e-5 3.43e-5 <dbl>       
#>  4 15-19  1993 Male        0 137470    17.5 2.84e-5 2.31e-5 3.43e-5 <dbl>       
#>  5 15-19  1994 Female      9 133010    17.5 2.69e-5 2.23e-5 3.24e-5 <dbl>       
#>  6 15-19  1994 Male        0 135920    17.5 2.69e-5 2.23e-5 3.24e-5 <dbl>       
#>  7 15-19  1995 Female      6 131960    17.5 2.70e-5 2.25e-5 3.27e-5 <dbl>       
#>  8 15-19  1995 Male        0 136430    17.5 2.70e-5 2.25e-5 3.27e-5 <dbl>       
#>  9 15-19  1996 Male        0 138000    17.5 2.63e-5 2.20e-5 3.17e-5 <dbl>       
#> 10 15-19  1996 Female      3 132490    17.5 2.63e-5 2.20e-5 3.17e-5 <dbl>       
#> # ℹ 650 more rows
#> # ℹ 1 more variable: .observed <dbl>