Extract components from a results object

Extract components from an object created by a call to smooth_age() or smooth_agetime().

Usage

# S3 method for class 'BayesRates_results'
components(object, what = NULL, interval = 0.95, ...)

Arguments

object

A results object.

what

Components to be extracted. If no value supplied, all available components are extracted.

interval

Width of credible intervals. A number between 0 and 1. Default is 0.95.

...

Not currently used.

Value

A named list of tibbles, or, if what is a single value, a single tibble. Each tibble contains

classification variables, if present

.fitted Point estimates (posterior medians).

.lower, .upper Lower and upper bounds of credible intervals.

.probability. A list column with all draws from the posterior distribution.

The age variable comes in versions: the original version and one with a .mid suffix. The .mid version is useful for plotting.

Details

If the object was created by a call to smooth_age(), then the components that can be extracted are:

"rates" Age-specific rates.

"intercept" Model intercept.

"age_effect" Age main effect.

"age_hyper" Hyper-parameters for age main effect.

If the object was created by a call to smooth_agetime(), then the components that can be extracted are

"rates" Age-time-specific rates

"intercept" Model intercept.

"age_effect" Age main effect.

"time_effect" Time main effect or age-time interaction.

"age_hyper" Hyper-parametres for age main effect.

"time_hyper" Hyper-parameters for time main effect or age-time interaction.

See also

Examples

## run model
results <- smooth_age(nevent_df = nz_divorces_2020,
                      py_df = nz_population_2020,
                      age_width_df = nz_age_width_df,
                      age_min = 15)

## extract all components
components(results)
#> $rates
#> # A tibble: 11 × 6
#>    age   age.mid   .fitted    .lower    .upper .probability 
#>    <fct>   <dbl>     <dbl>     <dbl>     <dbl> <list>       
#>  1 15-19    17.5 0.0000200 0.0000106 0.0000381 <dbl [1,000]>
#>  2 20-24    22.5 0.000493  0.000435  0.000560  <dbl [1,000]>
#>  3 25-29    27.5 0.00265   0.00254   0.00279   <dbl [1,000]>
#>  4 30-34    32.5 0.00522   0.00503   0.00542   <dbl [1,000]>
#>  5 35-39    37.5 0.00629   0.00606   0.00652   <dbl [1,000]>
#>  6 40-44    42.5 0.00702   0.00682   0.00724   <dbl [1,000]>
#>  7 45-49    47.5 0.00724   0.00702   0.00750   <dbl [1,000]>
#>  8 50-54    52.5 0.00638   0.00618   0.00657   <dbl [1,000]>
#>  9 55-59    57.5 0.00496   0.00476   0.00514   <dbl [1,000]>
#> 10 60-64    62.5 0.00327   0.00308   0.00348   <dbl [1,000]>
#> 11 65+      72.5 0.00121   0.00114   0.00130   <dbl [1,000]>
#> 
#> $intercept
#> # A tibble: 1 × 4
#>   .fitted .lower .upper .probability 
#>     <dbl>  <dbl>  <dbl> <list>       
#> 1   -6.12  -6.19  -6.06 <dbl [1,000]>
#> 
#> $age_effect
#> # A tibble: 11 × 6
#>    age   age.mid .fitted .lower .upper .probability 
#>    <fct>   <dbl>   <dbl>  <dbl>  <dbl> <list>       
#>  1 15-19    17.5  -4.70  -5.26  -4.11  <dbl [1,000]>
#>  2 20-24    22.5  -1.49  -1.61  -1.39  <dbl [1,000]>
#>  3 25-29    27.5   0.193  0.115  0.278 <dbl [1,000]>
#>  4 30-34    32.5   0.870  0.799  0.939 <dbl [1,000]>
#>  5 35-39    37.5   1.06   0.986  1.13  <dbl [1,000]>
#>  6 40-44    42.5   1.17   1.10   1.24  <dbl [1,000]>
#>  7 45-49    47.5   1.20   1.12   1.27  <dbl [1,000]>
#>  8 50-54    52.5   1.07   1.00   1.13  <dbl [1,000]>
#>  9 55-59    57.5   0.819  0.748  0.884 <dbl [1,000]>
#> 10 60-64    62.5   0.401  0.313  0.490 <dbl [1,000]>
#> 11 65+      72.5  -0.585 -0.674 -0.504 <dbl [1,000]>
#> 
#> $age_hyper
#> # A tibble: 2 × 5
#>   hyper .fitted .lower .upper .probability 
#>   <chr>   <dbl>  <dbl>  <dbl> <list>       
#> 1 sd       2.84  1.90    4.18 <dbl [1,000]>
#> 2 slope    1.10  0.728   1.48 <dbl [1,000]>
#> 

## extract age effect
components(results, what = "age_effect")
#> # A tibble: 11 × 6
#>    age   age.mid .fitted .lower .upper .probability 
#>    <fct>   <dbl>   <dbl>  <dbl>  <dbl> <list>       
#>  1 15-19    17.5  -4.70  -5.26  -4.11  <dbl [1,000]>
#>  2 20-24    22.5  -1.49  -1.61  -1.39  <dbl [1,000]>
#>  3 25-29    27.5   0.193  0.115  0.278 <dbl [1,000]>
#>  4 30-34    32.5   0.870  0.799  0.939 <dbl [1,000]>
#>  5 35-39    37.5   1.06   0.986  1.13  <dbl [1,000]>
#>  6 40-44    42.5   1.17   1.10   1.24  <dbl [1,000]>
#>  7 45-49    47.5   1.20   1.12   1.27  <dbl [1,000]>
#>  8 50-54    52.5   1.07   1.00   1.13  <dbl [1,000]>
#>  9 55-59    57.5   0.819  0.748  0.884 <dbl [1,000]>
#> 10 60-64    62.5   0.401  0.313  0.490 <dbl [1,000]>
#> 11 65+      72.5  -0.585 -0.674 -0.504 <dbl [1,000]>