Smooth rates over age and time

Use data on numbers of events and population at risk to produce estimates of rates that are smoothed over age and time. The amount of smoothing depends on the model used to represent age and time effects. Current choices for age effects are Spline() (the default) and RW2(). Current choices for time effects are TimeVarying() (the default) and TimeFixed(). See the documentation for Spline(), RW2(), TimeVarying(), and TimeFixed() for a description of the mathematical models underlying the smoothing.

Usage

smooth_agetime(
  nevent_df,
  py_df,
  agevar = "age",
  timevar = "time",
  byvar = character(),
  age_width_df = NULL,
  age_min = NULL,
  spec_age = Spline(),
  spec_time = TimeVarying()
)

Arguments

nevent_df

A data frame with a column called "nevent" containing the number of events, plus additional classification variables.

py_df

A data frame with a column called "py" containing exposure, plus additional classification variables.

agevar

The name of the age variable in nevent_df and py_df. Defaults to "age".

timevar

The name of the time variable in nevent_df and py_df. Defaults to "time".

byvar

The names of classification variables in nevent_df and py_df, other than age and time (eg "sex" or "region"). Optional.

age_width_df

A data frame with columns "age" and "width", giving the widths of each age group. Required if the age group labels are not integers.

age_min

The lower limit of the youngest age group. Required if the age group labels are not integers.

spec_age

The prior model for the age effect. Current choices: Spline() and RW2(). Defaults to Spline().

spec_time

The prior model for the time effect. Current choices: TimeFixed() and TimeVarying(). Defaults to TimeVarying().

Value

An object of class "BayesRates_results".

Details

Dataset nevent_df must contain a variable called nevent giving counts of events, and dataset py_df must contain a variable called py giving the population at risk.

nevent_df and py_df must also contain an age variable. If the age variable users multi-year age groups (eg "15-19" or "65+"), then smooth_age() needs help interpreting these age groups. A data frame age_width_df must be supplied giving the width of each age group, as well as a value age_min giving the lower limit of the youngest age group. (Future versions of BayesRates will try to parse age group labels automatically.)

nevent_df and py_df must contain a time variable. To analyse data without a time dimension, use function smooth_age(). Times must be single years, eg 2015, 2016, 2017. There may be gaps, eg 2010, 2015, 2020: smooth_agetime() automatically estimates values for the missing periods.

nevent_df and py_df can contain additional classifying variables, such as sex and region. Independent models are fitted for each combination of the levels of these additional variables.

See also

Examples

results <- smooth_agetime(nevent_df = nz_divorces,
                          py_df = nz_population,
                          age_width_df = nz_age_width_df,
                          age_min = 15)
results
#> --- Object of class "BayesRates_results" ---
#> 
#>      nevent ~ Poisson(rate * py)
#>   log(rate) = age_effect + time_effect
#>  age_effect ~ Spline()
#> time_effect ~ TimeVarying()
#> 
#>    agevar: age
#>   timevar: time
#>     byvar: <none>
#>    n_draw: 1000