Summarise rates by adding up over the life time

Construct a summary measure for age-specific rates by summing up the rates, weighted by the width of each age group. The calculations are equivalent to those used to construct the total fertility rate (TFR).

Usage

total_rate(x, interval = 0.95)

Arguments

x

An object of class "BayesRates_results", constructed via a call to smooth_age() or smooth_agetime().

interval

Width of credible intervals. A number between 0 and 1. Default is 0.95.

Value

A tibble with the following columns:

the time variable specified by timevar, if x was created by a call to smooth_agetime()

further classifying variables, as specified by byvar in the original call to smooth_age() or smooth_agetime()

.fitted. Point estimates from the model. (Posterior medians.)

.lower, .upper. Lower and upper limits of the credible interval specified by interval.

.probability. A list column containing all draws from the posterior distribution.

.observed. Total rates constructed from direct estimates, ie events divided by population, with no smoothing.

See also

Examples

res <- smooth_agetime(nevent_df = cn_divorces,
                      py_df = cn_population,
                      spec_time = TimeFixed())
total_rate(res)
#> # A tibble: 39 × 6
#>     time .fitted  .lower .upper .probability  .observed
#>    <int>   <dbl>   <dbl>  <dbl> <list>            <dbl>
#>  1  1980  0.0108 0.00599 0.0189 <dbl [1,000]>   0.00631
#>  2  1981  0.0120 0.00751 0.0196 <dbl [1,000]>   0.0133 
#>  3  1982  0.0124 0.00797 0.0191 <dbl [1,000]>   0.00749
#>  4  1983  0.0150 0.00967 0.0234 <dbl [1,000]>   0.0147 
#>  5  1984  0.0152 0.0102  0.0232 <dbl [1,000]>   0.00902
#>  6  1985  0.0181 0.0124  0.0266 <dbl [1,000]>   0.0248 
#>  7  1986  0.0175 0.0117  0.0256 <dbl [1,000]>   0.0154 
#>  8  1987  0.0178 0.0121  0.0260 <dbl [1,000]>   0.0116 
#>  9  1988  0.0226 0.0158  0.0316 <dbl [1,000]>   0.0220 
#> 10  1989  0.0253 0.0181  0.0353 <dbl [1,000]>   0.0177 
#> # ℹ 29 more rows