Specify a P-spline model for an age effect

Specify a P-spline (penalised spline) model for an age effect. A P-spline is flexible, but favours profiles that are relatively smooth.

Usage

Spline(df = NULL, scale = 1)

Arguments

df

Degrees of freedom for spline. The default is 0.7 times the number of age groups, or 4, whichever is higher.

scale

Scale for error term. The default is 1.

Value

An object of class "BayesRates_spec_spline".

Mathematical details

The model for the age effect, on the log scale, is

$$\beta = X \gamma$$

where

\(\beta\) is a vector of age effects,

\(X\) is a matrix holding the basis functions for the spline, with df columns, and

\(\gamma\) is a vector of coefficients, with df elements.

The elements of \(\gamma\) are assumed to follow a second order random walk,

$$(\gamma_a - \gamma_{a-1}) - (\gamma_{a-1} - \gamma_{a-2}) \sim \text{N}(0, \sigma^2)$$

Parameter \(\sigma\) has prior

$$\sigma \sim \text{N}^+(0, a^2)$$

where \(\text{N}^+\) denotes a half-normal distribution, and a value for \(a\) is supplied by the user.

See also

Examples

Spline()
#> < Object of class "BayesRates_spec_spline" >
#>           scale: 1
Spline(df = 10)
#> < Object of class "BayesRates_spec_spline" >
#>              df: 10
#>           scale: 1